[10126]: Move multierror into its own errors package, then rename it to just multi.
